[bindc] Add PCI library
Currently includes all of the (valid) PCI vendor IDs used in tree. The
library name includes "deprecated" because this is a temporary step
along the migration. The library currently does not define the
variables, instead it only extends variables in the "deprecated."
namespace. This is due to a limitation in the bind bytecode that will be
addressed once all drivers have migrated to using the bind compiler.

Change-Id: I2f1dccc7ac67c92d5f2f6dcd286eaac8c0263392