[debugger] Resolve virtual inheritance.
Adds a new ResolveInherited variant that can evaluate the asynchronous
location expressions needed for virtual inheritance.

Adds a test for virtual inheritance using real data from a test program.

Changes the DwarfExprEval to allow data to be pushed on the stack before
execution. This is required for the inheritance expressions.

This code is unused in production. Callers will be changed to the new
variant in followup patches.

Change-Id: I8855d559b10ca0747b7e26b58332bec29ebb081a