[fidl] Firm up the spec for fuchsia.intl.Profile
The specification was ambiguious as to whether the vector<> entries were
always filled in or not.  This spec confirms that the entries will
always have at least one element, and that there will always be a
default present.

This matters for general clarity, but also matters for languages like
rust where the entries must be Option<_> because of the backward
compatibility requirements that are unrelated to the actual needs of the
intl profile, but related to the way the FIDL API is implemented in
rust.

Change-Id: I08da91225dff763d474df5d8304a870c50f613c9