[debugger] Add additional general purpose registers
Supporting thread-local storage requires us to access the thread
pointer, which is offered in a register. Right now we don't support the
register needed on either arch.

This patch adds support for the necessary registers (fsbase on x64,
tpidr on arm), and throws in gsbase while we're at it for good measure.

Change-Id: I4370e060fb3239bdb497bc8c00d023e042a3ec20