[fdb] Add a FIDL service for starting a component.
We'll use this as the basis for the round 1 prototype using FDB to call
the Remote Control Service to start a component. I suspect there's a
better way to model this using handle's for stdout/stderr, but I'd like
to get this version in so we can do testing with gRPC/overnet asap.

Unfortunately I can't call the FIDL service RemoteControlService because
"service" is banned as a component of a service name.

Change-Id: If91a7c0362a41a53a64c8164a9522a55b61d12df