[debugger] Improve hex dumping.
Adds syntax highlighting to hex dumps.

Adds a mode to hex dump formatting to show offsets instead of absolute
addresses. This is useful when generating a hex domp for something that
doesn't have an intrinsic address. This is currently unused outside of
the test, I was originally going to use this to hex dump objects (rather
than pointers) but the command syntax is a bit confusing so am putting
this off.

Change-Id: I7767d13ce6744d5ff4adf69b7e4e2ee5c3133894