[wlan][iwlwifi] fixed a bug of unsafe 'calib_wait' usage.
This is a bug since fxr/296865 where line 601 skipped the removing of calib_wait
and caused the unsafe memory usage. This bug didn't show up until recently. It
started crashing from fxr/345952 where the symptom is memory access violation in
unrelated iwlwifi functions.

The fix is removing the 'calib_wait', which is a local variable, from the
mvm->notif_wait waiting queue.

Also add comments in notif-wait.h to point out how to use it correctly.

Tested: This CL passed the test code.
        It also makes fxr/345952 passed.
Change-Id: I5331f6e93048d23a016e0c963fc10bcb89f5c19e