[componentmgr] Replace Realm with Moniker in CapabilitySource
CapabilitySource currently holds onto a strong reference to a
Arc<Realm> (in the case of Component and Framework sources)

It turns out that all hooks listening to RouteCapability events
are only really interested in the moniker and not the realm itself.

If the hook wants certain guarantees on the Realm (it must be running
or resolved), it is the hook's responsibility to ensure those
guarantees by binding or looking up the realm using the moniker.

In fact, most hooks were already doing this, so passing the
Realm to them was pointless anyway.

The RootComponentResolved event has been generalized to
ResolveInstance. When any instance is resolved for the first
time, this event is dispatched.

This event allows WorkScheduler to listen for an instance resolution
and determine whether it exposes the Worker service to the
framework. Previously, this needed access to the Realm provided
by CapabilitySource.

Test: Ran all integration and unit tests that I could find.
Change-Id: I108f4500652702d5d2d5aaf25d0352ef5d50414f