[kernel][x86] Spectre V2: Track last aspace, avoid IBPB on kernel->user
Spectre V2 is a family of speculative execution attacks which target
indirect branch predictors/target address prediction.

Commit 024fbb3 modified the x86 thread switch code to issue an IBPB
(Indirect Branch Prediction Barrier) on every address space switch,
to mitigate Spectre V2 attacks by one user process against another.
Commit 6d8b144 improved the logic to skip the IBPB flush when switching
to a kernel thread as the kernel will protect itself via alternative
mechanisms (retpolines / IBRS+STIBP).

This commit skips the IBPB flush when switching from a kernel thread
back to a userspace one; the kernel is not expected to attack user
processes.

The combination of this optimization and the one in commit 6d8b144
needs care - if we switch from a user process to a kernel thread and
then to another user process, we will need to issue an IBPB flush; a
trivial implementation of both optimizations would fail to do so.

Instead track the last user address space that was active on a core
(lazily); when switching from the kernel to a new address space, then
issue an IBPB. This avoids a substantial fraction of IBPBs and
restores most of the performance lost by mitigating user/user Spectre
V2 attacks.

TODO(venkateshs): This CL has a small issue WRT address space reuse,
which I will address in a subsequent CL - if a user process switches to
a kernel thread, then exits, and a new process is created with the same
aspace*, and we switch to a thread in that process, we'll fail to issue
a flush.

Tested:
1) Booted on AMD 15h, added thread kcounters, compared IBPB counts:
   (before)
   $ kcounter | grep mp
   mp.ibpb = 93386 [382/sec]
   mp.thread_switches = 183273 [750/sec]

   (after)
   $ kcounter |  grep mp
   mp.ibpb = 85795 [111/sec]
   mp.thread_switches = 481625 [628/sec]

   ran 'k ut all', runtests.

2) Perfcompare -- this recovers the bulk of the performance lost by
   commit 16fa684 ("Flush Indirect Branch Predictor on process switch")
   on Fuchsia benchmarks, when there are a small number of runnable
   threads/processes:

   (enabling IBPB):
   RoundTrip_BasicChannel_SingleProcess 2.227-2.579  5189 +/- 298 ns  12417 +/- 197 ns
   (skipping IBPB flush on switching to kernel threads):
   RoundTrip_BasicChannel_SingleProcess 0.747-0.824  12413 +/- 118 ns  9745 +/- 382 ns
   (this CL):
   RoundTrip_BasicChannel_SingleProcess 0.473-0.603  9774 +/- 354 ns   5235 +/- 447 ns

Bug: 12540 Speculative Execution Mitigations.
Bug: 33667 Spectre mitigations?
Bug: 39621 Spectre V2: Don't issue Indirect Branch Predictor Barrier on every single aspace switch

Change-Id: I20a52e3339f42905f7c4b9f1ce56eddc02b3f354