[kernel][vm] Relax commited page requirement
Allows for changing the cache policy of a VMO even if it has committed
pages, provided none of the pages are pinned and that the VMO is
currently cached. When switching from a cached type any committed pages
are cleaned to ensure that any writes by the kernel (such as zeroing)
are visible.

Most usefully this change allows for changing the cache policy of a
contiguous VMO, as contiguous VMOs otherwise have their pages committed
on creation.

Change-Id: I64f8f04a2dcfdaec07be2ff9c4340fa22eb6c0b4