[bt][tools] Add pairing functionality to bt-cli
Adds functionality to call into the "Pair" function within the Bluetooth
control FIDL protocol. The CLI tool takes 3 arguments: a peer ID (which
autocompletes), a security level ("ENC" for encrypted or "AUTH" for
encrypted and authenticated), and a bondable mode flag ("T" or "F").
The device being paired to must already be connected for pairing to
succeed. This CL also restructures the "Expect" functions  in
ControlMock to use a macro to generate expect calls for different
ControlRequest enum variants.

Bug: 37650
Test: `fx run-test bt-cli-tests`, added tests "test_parse_pairing_\
      security_level", "test_parse_bondable_mode", "test_parse_pairing_\
      transport", "test_parse_pair", "test_pair" and "test_pair_error"
      to bt-cli/src/main.rs
Change-Id: I0cfeaf6d10e948a0a1ad0ee62699ebe517845887