[dart] Clean up handling of different main() interface variants
This wrapper code attemps to handle main() functions of two different
types:

  void main(List<String> args)
  void main()

Catching NoSuchMethodError is not the best way to handle this, because
that exception could get thrown somewhere in the body of main().  If
that happens, the wrapper will attempt to call main() a second time,
and the backtrace for the exception will be lost.

We can observe that with the following:

  void main(List<String> args) {
    print('here');  // This will get printed
    dynamic x = main;
    x.foo();  // This will throw NoSuchMethodError
    // After, the backtrace will show a failed zero-args main() call
  }

Instead we can do a runtime type check.

Bug: none
Test: "fx set terminal.x64", "fx run-host-tests sl4f_client_tests"
 - contains examples of 0-arg and 1-arg main() functions
   (see sdk/testing/sl4f/client/test)
Change-Id: I52eae9a39f9cf3563d628a9597e0731433d7dc42