[camera][gdc] Perform cache operations on configuration VMOs
After copying the GDC firmware configuration from a non-contig
VMO to a contig VMO, there is a need to clean and invalidate the cache
before the HW tries to read the configuration.

This CL enables the tests which serve as an integration test for
connecting all the supported streams and ensuring that streams are
flowing through the pipeline.

Fixed: 43029

Test: Enabled the disabled tests which were failing earlier.
      fx test camera_tests (x3)
      fx test arm_hermetic_camera_tests (x3)

Change-Id: Icffc1287ec6dd7d0b62c0482afb53ea83a95fb6f