[amlogic-decoder] Ensure clients don't allocate too many frames.
The h.264 decoder seems to have stringent requirements for the amount of
buffers - it must have at least the max number of reference frames + 6
(so often 22), but has 24 as max. Chromium was requesting to have 8
frames to use in the client, which meant sysmem was trying to allocate a
total of 30 frames. We can reduce the number of frames the client is
allowed to ask for, to ensure those can always be allocated.

Testability: new EXPECT in h264 integration tests + existing
use_h264_decoder_test.

Change-Id: I771202a4607b61cbafc94ad1fdf8c2bf7891939a