[utest] Remove flake in timers test.
We've seen a few flakes of the form:

    [ RUN      ] TimersTest.Basic
    ../../zircon/system/utest/timers/timers.cc:86: Failure: Expected timer.wait_one(ZX_TIMER_SIGNALED, deadline_wait, &pending) is ZX_OK.
        Expected: ZX_OK
        Which is: ZX_OK(0)
        Actual  : timer.wait_one(((zx_signals_t)1u << 3), deadline_wait, &pending)
        Which is: ZX_ERR_TIMED_OUT(-21)
    [  FAILED  ] TimersTest.Basic (5392 ms)

These flakes are QEMU (i.e., emulated) ARM systems, which run very
slowly and unpredictably w.r.t. the wall clock.

What appears to be happening is that we hit the 1 second syscall
deadline prior to the 50 millisecond timer (potentially on another QEMU
CPU) propagating to the kernel's internal event object.

Update the test timings to be able to deal with very slow machines.

Bug: 42005
Change-Id: Iacea8f10ca338bd9af7a54cb73e8eeddee6f5c8d