[fidl][rust] Verify padding is zero when decoding
This CL makes the Rust bindings use `skip_padding` instead of
`next_slice` to skip over padding, which ensures that the bytes are
indeed zero and produces a NonZeroPadding error if not. We also verify
that the entire inline portion of the buffer is consumed, and that
padding bytes after it (before the out-of-line portion starts) are zero.

This also fixes a hand-encoded FIDL buffer in dynlink.c, which was not
properly initialized to zero. This was breaking emulator startup with
the Rust padding checks enabled.

Test: fx run-test rust_fidl_tests
Fixed: 7922
Change-Id: Ia38c78c5a9207fbbd01d8bd7abe85997b75545b2