[futex][ownership] Fix a hypothetical cause of flake.
Make absolutely certain that our "waiter" thread is properly blocked
and waiting in our test futex before proceeding with the rest of this
test.  Failure to do this could introduce a race, which if lost can
result in tests failing when they should not because we cannot assign
ownership to a futex which currently has no waiters.

Bug: 42852
Change-Id: I7660c941c44ac2b6b8182fe3b9ce3eb07324f256