[msd-vsl-gc] Basic wait-link command queue implementation.
The command queue is represented by the ringbuffer, where we insert
new WAIT-LINK commands. When a command buffer is submitted, we modify
the last WAIT command in the ringbuffer to link to the command buffer.

We also need to link the command buffer back to the ringbuffer.

This is still missing some features. The ringbuffer will eventually
become full.

BUG=41095

Change-Id: I0c2d8933ffc27fac5d58e6c6660a30613e916ae3