[debugger] Make casting support callbacks.
Previously casting was always synchronous. The one asynchronous case was
when there was a reference that needs to be convered, and this was
handled by helper function and pushed onto the caller.

Virtual inheritance will require fetching memory so general casting must
now support more asynchronous evaluation. The reference case can now be
moved internal to the casting functions which cleans up the ExprNode
caller.

Assignment does some types of numeric casts which can still be handled
synchronously. A new function is added for these so the assignment code
can be kept synchronous.

There is no behavior change. The same casts are still supported in the
same way. Casting for virtual inheritance will be actually hooked up in
a followup.

Change-Id: I8a0289a5328df28e4f72978fa9d9e0d7c9bf5cb4