[netstack2] By default, install an on-link route for the IPv6 link-local subnet
Previously, a default route would be installed as an on-link route when a NIC's
state became connected. This would cause addresses on subnets that are not
actually on-link to be considered as such, resulting in IPv6-L2 address
resolution being attempted for nodes that are not actually neighbors, leading to
dropped packets.

Tests: Test that when a device's state becomes connected, an on-link route to
the link-local subnet is installed. Also test that when the state becomes down
or closed, the route is deleted.

Change-Id: Ice9c143a8666e9027ca2be28d706a9da44183acd