[zircon][syscalls] Add kernel.enable-serial-syscalls
This flag controls zx_debug_read() and zx_debug_write(). The default
value is false, meaning disabled, however the bringup product
configuration sets it to true meaning that it will be on in all
configurations where it is not explicitly turned off.

Also, update docs, and kernel command line documentation to spell out
which syscalls are disabled by which flags.

Bug: 38409
Change-Id: I9bcde67db4d297ec5ee49cfbec356cd1a2337479