[ext4] Basic parsing of ext4 image.
Can `ls` any given directory. Can read the data for any given file.

As a result, it can walk the filesystem and build a pseudo-fs-mt Simple
tree.

Supports the most basic ext4 filesystems only. Nested extents are not
supported, so only smaller and/or non-fragmented files are safe to read.
Does not check checksums, does not support 64bit filesystems, etc.

Bug: 37655
Test: fx run-test ext4-read-only-tests -t ext4_read_only_test
Change-Id: Ic219551bd1147879b049db10acd9073cfc967405