[codecs] Remove CodecFactory disconnect logging
The CodecFactory is designed to be a transient object that typically
only exists to connect a client to a Codec protocol. As such, client
errors are really of no consequence to the CodecFactory, so logging
disconnections from the client are not necessary. This CL removes the
logging of the client disconnect so as to not appear as an error.

Change-Id: I1efd2e95bacbdffde5cb45b816c73ea690c40c86