[sched][deadline] Map deadline profiles to priority
Map deadline profiles to a priority of 30 for systems which are not running with
the unified scheduler.  This behavior is meant to act as a bridge so that code
can specify a deadline profile and continue to operate (from an API standpoint),
even before the switch to a deadline aware scheduler has become the system's
default.

In the past, most threads which considered themselves to have realtime needs
have selected p24 as their priority.  This change now maps deadline profiles
them to p30 instead.

Change-Id: Icf2dbe9166f92d30fa056cb6ac073fbd496313af