[psutils] Fix use after free in "ps".
We additionally add some high-level smoke tests of basic "ps"
functionality, exercising enough of the code to expose the
use-after-free bug when run with ASAN and ensure it is fixed.

Finally, we remove global state from the ps implementation,
allowing us to call it multiple times from the test framework.

Thanks to jonathanbailey@ for the actual use-after-free diagnosis and
fix, included in this CL.

Fixed: 42417
Change-Id: Ifb9e92be2117c592ff5868e98699892019717baf