[debugger] Resolve vtable pointers.
Adds a function to convert a base class pointer or reference to a
derived class pointer or reference. This analyzes the vtable pointer to
see what object's vtable it's pointing into, and then automatically cast
the pointer to that type.

This code is not currently hooked up so there will be no user behavior
change.

Change-Id: I947f40001c03b1a26c7993d9b742467fdcb04919