[usb-virtual-bus][hid] Disable UsbHidTest.UnBind.
The UsbHidTest.UnBind test is triggering a heap-use-after-free error on
ASAN. I apologize for submitting a CL to disable the test without
communicating with the test owners first, but the test failure is
causing the bringup.x64-asan-qemu_kvm builder to fail for other CLs.
See the bug for the full ASAN report.

Bug: 43207
Test: This CL disables a test that was failing with ASAN.

Change-Id: I2d44a24eecdfe2eda8a9d2bdfc7bcf094051efad