[kernel][arm64][cache] fix the cache routines to properly iterate
In the last patch dealing with this code a small 'fix' was made to what
seemed like a typo in the ARM cache iterating code. Turns out the typo
was not a mistake but a subtle use of a 64bit vs 32bit register.

Fix the fix to the non-typo.

Found by booting on a raspberry pi 4 on qemu on kvm which has a 2 way
set associative L1d cache. All of the other cores we've booted on
previously have at least 3 way set association on every level. 2 way
caused the cache logic to get in a loop that it would never terminate
due to the way it computed the cpu flags in a 32bit subs instruction vs
64bit.

Also discovered by manually tracing through it that it was early
terminating on machines with > 2 way set association, so the routine
hasn't really been working right all along.

It's not used for much, only really as a safety clean before shutting a
cpu down and a safety clean as a cpu is booted, so its not likely to
change much except actually do what it says its going to do.

Tested: boot on all flavors of arm qemu and on a few pieces of arm
hardware with different set associativity.

Change-Id: I397298f5dc02bfb7528ad72da3b1feb042c8c790