[storage][directory_broker] Fix bug with empty relative paths
When an empty relative path is provided to a directory broker
that is created from a directory proxy, if the relative path
is empty, we used to clone with the same open flags.

However, since open flags are often unsupported by clone, the
right thing to do is to open with "." as the relative path.

This allows the proxy to handle what should happen.

If the proxy connects to a pseudo_fs::Simple, it will call
clone when it sees "." as the relative path anyway.

If the proxy connects to ServiceFs, then it can handle "."
appropriately, unlike the empty string.

This raises the interesting question: why is the empty string
considered an error anyway? Shouldn't it be equivalent to "."?

Updated hub integration test to use the Directory + POSIX flag
on open. DirectoryBroker used to do a clone causing ServiceFs to
return an error since these flags are not supported on clone.

Test: Re-ran hub integration test.
      Attempted to access /hub/exec/out for a v2 component.
      Failed at ServiceFs before this change. Works after.
Change-Id: I0bcad90b09d89aff276dc0bb02868f736da49654