[testsharder] Copy durations file into out dir
Adds a new "test_durations" build_api_module that copies a
test_durations.json file from a known location in integration.git to a
known (by testsharder) location at the root of the output directory.
This is necessary for testsharder to consume the file, since testsharder
is only aware of the location of the build dir and not the location of
the checkout dir.

Bug: 41674
Change-Id: I48fdb5889f6fd995d7e62bf02d745f81a8c9b596