[a11y] Add scrolling action to a11y FIDL.
While other accessibility features may support more complex scrolling actions, the screen reader merely needs to be able to request that an element (e.g. an item in a list) be scrolled to be visible onscreen. We may add additional scrolling actions in future.

Testing: It builds. No functional changes.

Change-Id: Id2cd9557d3b001b244b7b9e50e1c84c19f893151