[kernel][arm64] Add Main ID register/implementer to percpu structure
Some ARM64 code would benefit from being able to efficiently query
the specific core implementation it is running on (ex: Cortex-A53).
For example, speculative execution mitigations may only need to be
applied to Cortex-A57 cores in the A5* family.

Add an enumeration of core implementations and add it to the percpu
structure.

Tested:
* VIM2:
> ARM cpu 1: midr 0x410fd034 'ARM Cortex a-53 r0p4' mpidr 0x80000001 aff 0:0:0:1
* CN99XX:
> ARM cpu 0: midr 0x431f0af1 'Cavium CN99XX r1p1' mpidr 0x80000000 aff 0:0:0:0

Bug: 33667 Spectre mitigations?
Bug: 12540 Speculative Execution Mitigations.

Change-Id: I47e41b0b84e71827391e75ec756a26cad23a188f