[rust][fidl] Keep ClientInner ref until dropped
Previously EventReceiver would drop its reference to ClientInner
when the channel closed. Instead we now keep the reference, but
change the internal state of the EventReceiver.

Keeping the reference allows us to properly cleanup when the
EventReceiver is dropped, namely to reset the event receiver's
state on the ClientInner. This prevents a panic that occurred
previously when the channel closed, the EventReceiver dropped,
and someone tried to take the event stream. Now the event stream
can be taken again after being dropped by any other client.
The first call to poll_next from the new EventReceiver will
return None, subsequent calls to poll_next will panic, as they
did prior to this change.

Bug: 42902
Change-Id: I808aa19045d2afb976886d4e1e6793aa79965b83