[zircon][cobalt] Use ZX_ASSERT_MSG in MakeCobaltOptions
Use ZX_ASSERT_MSG instead of ZX_DEBUG_ASSERT_MSG in MakeCobaltOptions
to ensure that specifying an invalid project ID is fatal even when
built with GN assert_level = 0 (i.e. ZX_DEBUG_ASSERT is disabled).

This change is part of an effort to ensure the system works when GN
assert_level is 0.

Bug: 42977
Change-Id: Iba16fcacb672ba71c8031a7f03a0fb3a9d0d7628