[settings] Add short-term earcons settings service
This is a short-term solution to adding the earcons service. It is added
directly into the audio settings service. In the future, it should be
split out into its own service.

The earcons service right now supports two sounds, a volume changed and
a max volume sound. These can be heard by pressing the volume up/down buttons.
to include the sound player locally, add these flags to fx set:

--with //src/media/sounds --with-base //src/media/sounds/bundles:config

Fixed: 41405
Tests: The existing tests under setui_src/tests/audio_tests.rs cover the
check that the volume changes are made in the controller. There is a
TODO added for testing that the sounds are played or not based on the
logic. Because of the temporary placement of the code inside the audio
service, it is difficult to mock the dependency.

fx run-test setui_service_tests -- --test
(Also manually tested)

Change-Id: If9d9ba0e03e3c2e291e62a055170d9243a1dfa3c