[kernel] Make sancov collect counts for coverage points
This changes the sancov instrumentation in the kernel to count how
many times each coverage point is executed.  It now delivers two
"live data" VMOs: sancov and sancov-counts.  The former is as
before, in the upstream-compatible format of a simple list of PCs.
The latter is a similarly simple format that goes along with that:
each word in the sancov-counts VMO is a counter for the coverage
point whose PC is at the same index in the sancov (PCs) VMO.

Bug: 34197
Test: netcp my-machine-name:/boot/kernel/data/zircon.elf.1.sancov-counts sancov-counts && od -xt x8 sancov-counts
Change-Id: If2d349775dba83c5d298988e11096d67e622565a