[futex] Fix flake when using sync_mutex with BAD_HANDLE policy
When a futex is under contention, there's a race between

https://fuchsia.googlesource.com/fuchsia/+/08ad01f521b1523b29fbacf281f3ce3317ef9063/zircon/system/ulib/sync/mutex.c#28
and entering _zx_futex_wait() on the following line

vs.

the thread named in old_state calling sync_mutex_unlock() (and zeroing
out its thread handle).

- Thread A trying to lock in the slow path, gets past the atomic and
  loads old_state (which contains thread Bs handle).
- Then thread A yields to thread B.
- Thread B unlocks the mutex and exits (which closes the thread B handle).
- Then thread A resumes and passes thread B's handle into
  _zx_futex_wait() to assign ownership.
- With the BAD_HANDLE job policy, this will cause the caller to be
  killed through no fault of its own.

This was rarely happening in reboot_tests at boot when devmgr added a
bunch of devices on helper threads using a sync_mutex to guard.

Test: deterministic new case added to futex-ownership-test
Bug: 41780
Change-Id: I97d9ad007aafe08b95bc8962e0e51dc890f84f79