[a11y] Remove cyclic dependency in gesture_manager
Currently, gesture_handler and gesture_arena has a cyclic dependency.
This can result in a crash. Refer to the bug for more context.

This change removes gesture_arena dependency from gesture handler and
reverses the order of initialization of gesture_handler and arena in
gesture manager so that gesture handler gets deleted before gesture
arena.

Test:
  * Build
  * fx run-test a11y_lib_tests
  * fx run-test a11y_tests
  * Manually verified that test doesn't crash anymore.

BUG=43187

Change-Id: Ifead7218be594e19ce74bde11f9718b41490eb12