[debugger] Automatically convert to derived classes.
Adds a new option on the EvalContext that controls whether base classes
are automatically converted to derived classes when possible. The
currently always returns false in production so there should be no
user-noticeable change. This will be hooked to the settings system in a
future pass.

Hooks up the flag to reference resolution. If the flag is set to true,
references will be automatically converted to derived classes. This will
enable "print foo.bar" to work with "Base& foo".

Pointers and printing are still not hooked up.

Separate out the derived virtual class setup from resolve_base_unittest
to a separate file so it can be shared by other tests. Use this in the
new reference resolution test code.

Change-Id: I4cd4b1e5e07cd640ae1528cfc9edc745191fa0a8