[docs][components] Remove references to the implicit ELF runner.
The ELF runner is no longer a capability that is provided by default.
Instead, it must now be explicitly used and routed.

Update prose in components v2 documentation to remove claims that
the ELF runner will be provided by default, and update all CML examples
with a `program` block to also `use` the ELF runner capability (and
route to children, where appropriate).

Bug: 4761
Change-Id: If39b771f7a592e64c801f84ab19e4640c2e55b66