[carnelian] Use ImagePipe2 for scenic canvas mode
This CL modifies Carnelian to use ImagePipe2 as an implementation
details of Scenic canvas mode.

The reason for this change is to align the implementations of Carnelian
apps running under Scenic and directly on the display controller.
Currently these modes are not sharing much, but this change will allow
a future change to combine a lot of the implementation.

Using ImagePipe2 added a number of asynchronous operations to
previously synchronous parts of Carnelian's implementation. The bulk of
this CL are changes to allow the use of Rust's async/await to simplify
these asynchronous operations.

The biggest change was the removal of the App::with method that
previously was the way to communicate between async callbacks and
various parts of a Carnelian app. This has been replaced with message
passing using Fuchsia-Async channels. The sending half of the channel
can be cloned, so it can be handed off to any async callback that needs
it.

This CL also contains a related change that makes requires apps to
provide a future to create the app assistant, rather than providing it
directly. This allows the creation of app assistants to include
async/await calls as needed for initialization. Something similar will
likely happen for view creation in a future CL.

Testing: Manual testing, existing unit and integration tests along with
a new unit tests for sysmem allocation.

Change-Id: I2ab9f2d92d4a8fe31d277607b4ff3dd03e58e20d