[ledger] Handle inline empty tree node in BTree builder
A different version of the Flatbuffer library produces a more compact
encoding of the empty Treenode, which makes PageStorageTest check-fail
because MarkAllPiecesLocal expects non-inline new objects. This commit
updates the BTree builder to uphold this expectation, guarding against
an inline node before adding the empty tree node identifier to the list
of new objects.

We hope really really hard that non-empty tree nodes will never be
inlined, because they will contain the object identifier of at least one
child, or at least one entry and the overhead of Flatbuffers would be
large enough even for a very very tiny entry. But because we are
paranoid at this point, we add a DCHECK to catch potential violations of
this invariant more early and systematically.

Tested: PageStorageTest does not check-fail anymore with the new Flatbuffer version.
Change-Id: I18c39ffd0c543f93fef0cddaaf2a16187ecdc37a