[wlan][iwlwifi] Tx cmd supports NOCOPY and DUP flags.
Both flags now do the same thing: copy the data into an io_buffer.

For DUP: Original code uses a free_buf pointer. But it cannot be mapped
         into a physical address. Need an io_buffer to map it.

For NOCOPY: The flag was designed for not copying large packet due to
            performance issue. However, same as DUP, we cannot map the
            virtual address to physical address. So, copy it.

It turns out that we merge them into the same behavior: copy data to
a io_buffer. Then release it in Rx interrupt.

Bug: 37782
Tested: Added test cases to cover it.

Change-Id: I6da1e7f45dd9389bdc0829940055ba7e2af0b8d7