[fuchsia-async] don't re-use wait_async port keys
If a ReceiverRegistration is dropped, it's removed from the receiver
map, but nothing is presently done to cancel incoming packets for its
corresponding key. So if we re-use the key for a different purpose, we
might route that pending packet to the wrong receiver.

Ideally we'd figure out how to properly cancel the incoming packets, but
there's some subtle nuances to doing so safely in a multithreaded
executor.

So for now, we'll just not reuse keys, and hope that there are a lot of
u64s.

Bug: 43101
Change-Id: I3a33ceab8c0d219bbd0e1ae4386d6f550141a0f1