[SetUI] Terminate hanging get handlers.
Currently, hanging get handlers spawn an async thread to process
updates. This thread persists even after the channel is closed. As
a result, the thread lives on indefinitely, leading to an increasing
memory usage.

This changelist corrects this issue by explicitly signaling to the
receiving thread to close upon channel exit and when all
references to the hanging get handler are lost.

Fixes: 42254
Test: fx run-test setui_service_tests -- --test
Change-Id: If4dd7fce30b36eebac179829579293d21545dd49