[debugger] Use the platform message loop.
The debug_agent hardcoded the "target" message loop. Use the platform
message loop which changes its type based on the current platform.

Add a way for the message loop's Init function to return success/failure
and an error message. Update callers.

The platform message loop had bitrotted on Fuchsia when we renamed the
class, since it was not used. Fix the definition.

Convert zxdb callers that had manually instantiated a message loop to
use the TestWithLoop helper which automatically sets things up.

There is no behavior change.

Change-Id: Ib94571c9c5fc45a83a972a005d4b0ea3bfd7cc58