[debugger] Separate out some arch functions.
Previously the ArchProvider was mostly a container that allowed for a
few functions to be overridden for mocks.

This change is a first step to make it a virtual interface. The
architecture-independent platform parts are implemented by
ArchProviderFuchsia, and the architecture-specific parts are implemented
by the X64 and Arm64-specific subclasses of that.

A typedef is added for the cases that want the current platform's
concrete provider type (ArchProviderImpl), and the MockArchProvider is
updated. Some tests ahd specific mocks, and these now inherit from the
shared MockArchProvider.

There is no behavior change.

Change-Id: Ief9224045e803304444185b7a19f4e2a972f46f8