[benchmarks] Declare minimal set of package deps needed for peridot-perf builder
This will allow the peridot-perf builder to be switched to building
the same set of packages as terminal-x64-release --
i.e. //products/terminal.gni + //bundles/buildbot:terminal instead of
the current //products/core.gni + //bundles/buildbot:core.

That will reduce the build time of peridot-perf from ~33 minutes to
~13 minutes.

This is a stop-gap to reduce resource usage on the CQ until the SL4F
test wrappers on the 'terminal' builder replace the peridot-perf
builder.

Do this by adding the remaining perf test dependencies to be included
(indirectly) by terminal.gni.

Change the zircon_benchmarks invocations so that they only depend on
zircon_benchmarks_perftestmode and not on
zircon_benchmarks_unittestmode.  (This matches the other invocation in
src/tests/end_to_end/perf/test/microbenchmarks_test.dart.)

Steps for testing that switching peridot-perf to use terminal.gni
would work after this CL:

 * On infra: Test peridot-perf trybot via 'led' tool:

   From "recipes" dir in an infra checkout:

   led=$(pwd)/../prebuilt/tools/led
   $led get-builder 'luci.fuchsia.try:peridot-x64-perf-dawson_canyon' > led_plan
   edit led_plan to change "core" to "terminal" (see above)
   # Apply this change:
   $led edit-cr-cl https://fuchsia-review.googlesource.com/c/fuchsia/+/347581 < led_plan > led_plan2
   # Work around part of fuchsia_perf.py that fails when run via 'led':
   edit recipes/fuchsia_perf.py to disable api.buildbucket.get_build() call
   (PATH=$(pwd)/../prebuilt/tools/:$PATH $led edit-recipe-bundle) < led_plan2 > led_plan3
   $led launch < led_plan3

 * Local:

   fx set terminal.x64
   build + boot etc.
   fx shell 'mkdir -p /tmp/infra-test-output && /pkgfs/packages/fuchsia_benchmarks/0/bin/benchmarks.sh /tmp/infra-test-output --catapult-converter-args --bots dummy_bots --masters dummy_masters --execution-timestamp-ms 1 --log-url dummy_url'

   In order to match what infra does, don't run "fx serve-updates".
   This means that all Fuchsia packages required by tests must be in
   base_package_labels or cache_package_labels.  Packages in
   universe_package_labels don't get made available.

Bug: 42288
Test: on CQ: peridot-perf, terminal-x64-release
  + semi-manual steps above
Change-Id: I2031ae2dcea6bbe71ab9e8486524296b63574842