[scenic] FlatlandRenderer V1
This CL creates a renderer called RectangleCompositor,
which is optimized to efficiently render batches of 2D
rectangles, without the overhead of PaperRenderer. This
renderer will be a subsystem used within the larger
Flatland Renderer.

This version is able to take in vectors of FlatlandDrawables,
each of which represent a rectangle, color, and texture,
and render then to an output image while taking into
account transparency.

Rotations that are multiples of 90 degrees are also handled
implicitly by setting the proper UV coordinates on the
FlatlandSourceSpec struct.

Testing: New unit tests in rectangle_compositor_unittest.cc.
These tests verify that all functionality such as
transparency, rotation, uv-texturing, etc are working
as expected.

Change-Id: I80568a75be763ce1fc1b8c75a196a23dad85d7c4