[fdio,zxio] Defer to zxio in fdio_zxio_remote_wait_{begin,end}
Previously fdio_zxio_remote_wait_* translates POSIX events directly to
Zircon signals, with some bugs (for example, the WRITABLE signal for
files and devices are different, but the code uses the signal values
from devices for both). This CL changes that to call into zxio, and
introduces different treatment for signal translation between files and
devices.

Change-Id: Ie7e0a6de1889a4327b0b0b4630f762a410906407