[fidl][rust] Read both hashed and explicit xunion ordinals
This updates the rust bindings to accept the two sets of
ordinals specified in the "hashed_ordinal" and "explicit_ordinal"
fields when decoding xunions.
The value in "ordinal" is still used when encoding and as the
return value of `.ordinal()` - this is currently the hashed
ordinal for most xunions, and the explicit ordinal for static
unions that are now xunions, but will move towards being
the explicit ordinal for all xunions once read support for
both sets has landed across all bindings.

Test: fx run-test rust_fidl_tests
Change-Id: Ic99c81bb05dd92b3557d4743d7962ae86290ae43