[bindc] Add ACPI bind library
The library name includes "deprecated" because this is a temporary step
along the migration. The library currently does not define the
variables, instead it only extends variables in the "deprecated."
namespace. This is due to a limitation in the bind bytecode that will be
addressed once all drivers have migrated to using the bind compiler.

Change-Id: Id2d2400c3f517dd187871b0e04a0641f19d2b41b