[bindc] Add HID bind library
The library name includes "deprecated" because this is a temporary step
along the migration. The library currently does not define the
variables, instead it only extends variables in the "deprecated."
namespace. This is due to a limitation in the bind bytecode that will be
addressed once all drivers have migrated to using the bind compiler.

Change-Id: I9cb6d419c35e2e7d22784ca5405241fcfa3a9359