[build] Add test_durations to build modules
The build system now copies the test_durations.json file into the root
of the out dir as a build API module. It will eventually be used by
testsharder to construct shards with roughly even expected times. But
before we can do that, the file needs to be read in as a new module.

I added a test to ensure that the raw milliseconds are correctly
deserialized to time.Duration values.

Bug: 41674
Change-Id: I4fd2fc79ac0c985990780675cb564909e2837b5c