[testsharder] Produce test shards of ~equal duration
Uses the test_durations.json file copied from integration by the build
system to divide tests into shards such that each shard has
approximately the same expected total duration, rather than the same
number of tests.

This will hopefully speed builds up by a few minutes
once the durations file is seeded with actual data; for now, it applies
the same duration to every test, so this change alone won't have any
effect.

It's somewhat tricky to test the new sharding algorithm without relying
heavily on implementation details such as the ordering of the output
shards and the distribution between shards of tests of the same
duration. I tried to minimize the dependency on implementation details
by creating the assertShardsContainTests helper function, which ignores
the ordering of shards and the ordering of tests within each shard.

Bug: 36480
Change-Id: Iac78e04c1447e662a6884dca1c49708a0fd780a7