[pm] publish: add a flag to prevent creating repos if they don't exist
A common SDK use case is to `pm publish` a repo, but the current
behavior of `publish` creates a repo in a directory if it doesn't
already find one there. This is not always desirable if the user made
a typo, or if they've previously asserted that a repo should exist at
a given location.

With the `-n` flag specified, if a repo does not exist at a given
location, it will not be created, and `pm publish` will exit with an error.

Fixed: 43159
Change-Id: I23ec1e2e6bfee72521cdbe76762ddebcf80a9ded