[fidlc] Restructure fidlc json golden workflow
This CL updates the way JSON golden files are generated
and tested in fidlc.

Instead of being described inline as strings inside
`json_generator_tests.cc`, the test fidl files are moved
to a `typestest/` directory. Similar to fidlgen, this
directory also contains a `regen.sh` script that will run
fidlc on all inputs in `typestest/` and output the json to
`goldens` directory. This script has the added functionality
of handling fidl libraries with dependencies correctly.

As before, `json_generator_tests.cc` checks that the current
state of fidlc matches the goldens but does so now by
running fidlc on each file inside `typestest/` and comparing it
to the expected output in `goldens/`. This is implemented by
creating a goldens.cc file at build time containing all
of the goldens data, similar to the other fidl-compiler-tests.

Fixed: 38595

Test: manual
Change-Id: I268965f382b5e17dac7261c3705ec614ac5e2a4d