[identity] Use OauthOpenIDConnect protocol for id token requests
This switches token manager to use the new protocol
when requesting an id token.  Obtaining user info
is still done through the AuthProvider protocol
as we still need it to walk through the steps
of obtaining a refresh token.
This also deletes the GetAppIdToken call from
the AuthProvider protocol as it is no longer used.

Testing: unit tests, token manager integration tests

Bug: 565

Change-Id: Ibc62ce91f2bc12d9d80b4edb0b319ed409dd7e53