[cobalt][lib] fix two bugs
* The callback passed to CreateLogger*() would never execute
because logger_factory would go out of scope at the end of
ConnectToCobaltApplication() and close its underlying channel.
BaseCobaltLoggerImpl now owns logger_factory_ and the channel
is unbound at the end of the callback.

* The above change exposed a new bug in BaseCobaltLoggerImpl and was
reverted in (fxr/349971).

* This happened becuase of the way BaseCobaltLoggerImpl determines
whether it can send events to the remote Logger. BaseCobaltLoggerImpl
initially intended to only send events to the remote Logger if and only
if all sent events had been acknowledged positively and the remote
Logger was ready. The latter was determined by checking if the
InterfacePtr for the Logger (logger_) was bound, but logger_ was
always bound to a channel regardless of whether the remote Logger was
ready. This was due to the fact that a request is to logger_ in
ConnectToCobaltApplication() by each of the CreateLogger*() functions.

* The bug triggered with following flow:
	1) Create an instance of BaseCobaltLoggerImpl. This binds
	logger_ and sends a request to create a Logger to the remote
	LoggerFactory.
	2) Before the callback logger_factory_->CreateLogger*()
	executes, log multiple events with BaseCobaltLoggerImpl. The
	first logged event will be determined sendable in LogEventOnMainThread
	becuase logger_ is bound and there are no events in transit.
	This event will be marked as in transit and sent to the remote
	Logger. The callback that acknowledges that event was received
	successfully will later remove it from the set of events in transit.
	The rest of the events logged in this sequence will not be considered
	sendable since the set of events in transit is not empty.
	3) The callback in logger_factory_->CreateLogger*() executes,
	calling SendEvents(). At this point a deug build will fail the
	DCHECK since the first logged event is in the set of events in
	transit. If in a release build, the first event logged will be
	erased becuase of the use of std::move in SendEvents().
	4) The callback that cleans up the first event executes since
	the remote Logger is now set up by the LoggerFactory. The
	callback tries to remove the first event from the set of events
	in transit, however the first event no longer exists and the
	program using the library crashes.

TESTED=`fx run-test cobalt_lib_tests -t cobalt_lib_unittests`
TESTED=manually verified log messages stating channel was closed stopped
TESTED=manually verified basemgr does not crash on boot

Change-Id: I02828244d665dc85e6f25ce471d54861a925b6bf